+++ Examples/AdMobRewardedAdapterExample/app/src/main/java/com/yandex/admobadapter/rewarded/sample/MainActivity.java
@@ -9,17 +9,19 @@
 package com.yandex.admobadapter.rewarded.sample;
 
 import android.os.Bundle;
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
 import com.google.android.gms.ads.AdRequest;
-import com.google.android.gms.ads.MobileAds;
-import com.google.android.gms.ads.reward.RewardItem;
-import com.google.android.gms.ads.reward.RewardedVideoAd;
-import com.google.android.gms.ads.reward.RewardedVideoAdListener;
+import com.google.android.gms.ads.LoadAdError;
+import com.google.android.gms.ads.OnUserEarnedRewardListener;
+import com.google.android.gms.ads.rewarded.RewardItem;
+import com.google.android.gms.ads.rewarded.RewardedAd;
+import com.google.android.gms.ads.rewarded.RewardedAdLoadCallback;
 
 import java.util.Locale;
 
@@ -28,33 +30,14 @@ public class MainActivity extends AppCompatActivity {
     private static final String AD_UNIT_ID = "ca-app-pub-XXXXXXXXXXXXXXXX/YYYYYYYYYY";
 
     private Button mLoadAdButton;
-    private RewardedVideoAd mRewardedVideoAd;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        mLoadAdButton = (Button) findViewById(R.id.load_button);
+        mLoadAdButton = findViewById(R.id.load_button);
         mLoadAdButton.setOnClickListener(new RewardedAdClickListener());
-    }
-
-    @Override
-    protected void onDestroy() {
-        if (mRewardedVideoAd != null) {
-            mRewardedVideoAd.setRewardedVideoAdListener(null);
-        }
-        super.onDestroy();
-    }
-
-    @NonNull
-    private RewardedVideoAd createRewardedAd() {
-        final RewardedVideoAd rewardedVideoAd = MobileAds.getRewardedVideoAdInstance(this);
-
-        final RewardedAdListener rewardedAdListener = new RewardedAdListener();
-        rewardedVideoAd.setRewardedVideoAdListener(rewardedAdListener);
-
-        return rewardedVideoAd;
     }
 
     private class RewardedAdClickListener implements View.OnClickListener {
@@ -63,37 +46,32 @@ public class MainActivity extends AppCompatActivity {
             mLoadAdButton.setEnabled(false);
             mLoadAdButton.setText(getResources().getText(R.string.start_load_ad_button));
 
-            mRewardedVideoAd = createRewardedAd();
-
             /*
              Replace AD_UNIT_ID with your unique Ad Unit ID.
              Please, read official documentation how to obtain one: https://apps.admob.co
              */
-            mRewardedVideoAd.loadAd(AD_UNIT_ID, new AdRequest.Builder().build());
+            final AdRequest adRequest = new AdRequest.Builder().build();
+            final RewardedAdListener rewardedAdListener = new RewardedAdListener();
+            RewardedAd.load(MainActivity.this, AD_UNIT_ID, adRequest, rewardedAdListener);
         }
-    };
+    }
 
-    private class RewardedAdListener implements RewardedVideoAdListener {
+    private class RewardedAdListener extends RewardedAdLoadCallback implements OnUserEarnedRewardListener {
 
         @Override
-        public void onRewardedVideoAdLoaded() {
-            if (mRewardedVideoAd != null) {
-                mRewardedVideoAd.show();
-            }
+        public void onAdLoaded(@NonNull final RewardedAd rewardedAd) {
+            super.onAdLoaded(rewardedAd);
+
+            rewardedAd.show(MainActivity.this, this);
 
             mLoadAdButton.setEnabled(true);
             mLoadAdButton.setText(getResources().getText(R.string.load_button));
         }
 
         @Override
-        public void onRewarded(final RewardItem rewardItem) {
-            final String rewardedCallbackMessage = String.format(Locale.US, "onRewarded(), %d %s",
-                    rewardItem.getAmount(), rewardItem.getType());
-            Toast.makeText(MainActivity.this, rewardedCallbackMessage, Toast.LENGTH_SHORT).show();
-        }
+        public void onAdFailedToLoad(@NonNull final LoadAdError loadAdError) {
+            super.onAdFailedToLoad(loadAdError);
 
-        @Override
-        public void onRewardedVideoAdFailedToLoad(final int errorCode) {
             mLoadAdButton.setEnabled(true);
             mLoadAdButton.setText(getResources().getText(R.string.load_button));
 
@@ -101,18 +79,10 @@ public class MainActivity extends AppCompatActivity {
         }
 
         @Override
-        public void onRewardedVideoAdOpened() {}
-
-        @Override
-        public void onRewardedVideoStarted() {}
-
-        @Override
-        public void onRewardedVideoAdClosed() {}
-
-        @Override
-        public void onRewardedVideoAdLeftApplication() {}
-
-        @Override
-        public void onRewardedVideoCompleted() {}
-    };
+        public void onUserEarnedReward(@NonNull final RewardItem rewardItem) {
+            final String rewardedCallbackMessage = String.format(Locale.US, "onRewarded(), %d %s",
+                    rewardItem.getAmount(), rewardItem.getType());
+            Toast.makeText(MainActivity.this, rewardedCallbackMessage, Toast.LENGTH_SHORT).show();
+        }
+    }
 }