+++ SDK/adapter/mopub-mobileads/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to Yandex Adapter for MoPub Mediation will be documented in this file.
 
+## Version 0.5.0
+
+#### Added
+* Added support for Yandex Mobile Ads SDK version 4.1.0
+* Updated minimum supported Yandex Mobile Ads SDK version to 4.1.0
+* Improvements and optimizations
+
 ## Version 0.4.0
 
 #### Added